5.0.0.10302013
Developer's note: It's been a hectic release schedule this past month, and from this point on the, the updates are going to slow down tremendously (unless code breaking bugs are reported or found).  This plugin has grown in the past month to the monster it is now - and as the developer, I'd like to thank you for your continued use of it.  

Exclude installation bug fix.

Further code cleanup on remaining files that weren't previously cleaned up in the 4.0.+ series.

Install procedures (activate all, deactivate all, reset, nuke) corrected.

Module settings pages have had their CSS altered to be more uniform.


git-svn-id: https://plugins.svn.wordpress.org/my-optional-modules/trunk@796252 b8457f37-d9ea-0310-8a92-e5e31aec5664